django-project-main/requirements.txt
Django>=4.2.0
django-allauth>=0.54.0
channels>=4.0.0
channels-redis>=4.1.0
stripe>=7.0.0
django-htmx>=1.17.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-dotenv>=1.0.0
django-environ>=0.11.2
redis>=5.0.0
googlemaps>=4.10.0
django-filter>=23.0
django-storages>=1.14
whitenoise>=6.6.0 